apps/auth-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
python-multipart
passlib[bcrypt]==1.7.4
bcrypt==3.2.0